colab_demo/services/requirements-colab.txt
fastapi
uvicorn
python-dotenv
python-multipart
requests
numpy
Pillow
opencv-python-headless
scikit-image
ultralytics>=8.3.0
git+https://github.com/ultralytics/CLIP.git
vtracer
pypotrace
gradio
pytest
huggingface_hub
